backend/billing-service/requirements.txt
fastapi[standard]
uvicorn
google-cloud-pubsub
python-dotenv